# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A CRUD server for analyzing financial portfolios comprising stocks, bonds, and ETFs.",
        "title": "Portfolio API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/users/{user_id}/glance": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "glance"
                ],
                "summary": "List pinned portfolios with key metrics",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pre-IPO gap strategy: empty (constrain start date), cash_flat, cash_appreciating, or reallocate",
                        "name": "missing_data_strategy",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GlanceListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "glance"
                ],
                "summary": "Pin a portfolio to the user's glance list",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Portfolio to pin",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddGlanceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{user_id}/glance/{portfolio_id}": {
            "delete": {
                "tags": [
                    "glance"
                ],
                "summary": "Unpin a portfolio from the user's glance list",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Portfolio ID",
                        "name": "portfolio_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.AddGlanceRequest": {
            "type": "object",
            "required": [
                "portfolio_id"
            ],
            "properties": {
                "portfolio_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.GlanceListResponse": {
            "type": "object",
            "properties": {
                "portfolios": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GlancePortfolio"
                    }
                }
            }
        },
        "models.GlancePortfolio": {
            "type": "object",
            "properties": {
                "current_value": {
                    "type": "number"
                },
                "daily_return": {
                    "$ref": "#/definitions/models.ReturnMetric"
                },
                "life_of_portfolio_return": {
                    "$ref": "#/definitions/models.ReturnMetric"
                },
                "name": {
                    "type": "string"
                },
                "one_month_return": {
                    "$ref": "#/definitions/models.ReturnMetric"
                },
                "one_year_return": {
                    "$ref": "#/definitions/models.ReturnMetric"
                },
                "portfolio_id": {
                    "type": "integer"
                },
                "valuation_date": {
                    "type": "string"
                },
                "warnings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Warning"
                    }
                }
            }
        },
...
        "models.ReturnMetric": {
            "type": "object",
            "properties": {
                "dollar": {
                    "type": "number"
                },
                "percentage": {
                    "type": "number"
                },
                "start_date": {
                    "type": "string"
                }
            }
        },
...
        "models.Warning": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/models.WarningCode"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.WarningCode": {
            "type": "string",
            "enum": [
                "W1001",
                "W1002",
                "W1003",
                "W2001",
                "W3001",
                "W3002",
                "W4001",
                "W4002",
                "W4003",
                "W4004"
            ],
            "x-enum-comments": {
                "WarnBenchmarkDataUnavailable": "benchmark ticker missing or has no price data; Alpha/Beta set to zero",
                "WarnCashSubstituted": "pre-IPO period covered with synthetic cash prices; start date unchanged",
                "WarnETFSourceIncomplete": "source ETF data does not add up to 100%",
                "WarnExcessiveForwardFill": "too many securities needed forward-filling on some dates; those dates excluded",
                "WarnFuzzyMatchSubstituted": "dash-inserted ticker used in place of original (e.g. BRKB → BRK-B)",
                "WarnMissingPriceHistory": "one or more securities have no price history; affected dates excluded",
                "WarnPartialETFExpansion": "holdings scaled to 100% because resolved weights didn't sum to 1.0",
                "WarnProportionalReallocation": "pre-IPO period handled by redistributing weight among available members",
                "WarnStartDateAdjusted": "start date adjusted to security inception date",
                "WarnUnresolvedETFHolding": "individual unresolved holding (dropped from results)"
            },
            "x-enum-descriptions": [
                "individual unresolved holding (dropped from results)",
                "holdings scaled to 100% because resolved weights didn't sum to 1.0",
                "source ETF data does not add up to 100%",
                "dash-inserted ticker used in place of original (e.g. BRKB → BRK-B)",
                "one or more securities have no price history; affected dates excluded",
                "too many securities needed forward-filling on some dates; those dates excluded",
                "start date adjusted to security inception date",
                "benchmark ticker missing or has no price data; Alpha/Beta set to zero",
                "pre-IPO period covered with synthetic cash prices; start date unchanged",
                "pre-IPO period handled by redistributing weight among available members"
            ],
            "x-enum-varnames": [
                "WarnUnresolvedETFHolding",
                "WarnPartialETFExpansion",
                "WarnETFSourceIncomplete",
                "WarnFuzzyMatchSubstituted",
                "WarnMissingPriceHistory",
                "WarnExcessiveForwardFill",
                "WarnStartDateAdjusted",
                "WarnBenchmarkDataUnavailable",
                "WarnCashSubstituted",
                "WarnProportionalReallocation"
            ]
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}